# docs/swagger.yaml
basePath: /api/v1
definitions:
  database.OSSConfig:
    properties:
      access_key:
        description: 访问密钥ID
        type: string
      auto_sync:
        description: 是否开启自动同步
        type: boolean
      bucket:
        description: 存储桶名称
        type: string
      created_at:
        description: 创建时间
        type: string
      endpoint:
        description: 自定义端点
        type: string
      id:
        type: integer
      is_active:
        description: 是否为当前激活的配置
        type: boolean
      is_enabled:
        description: 是否启用
        type: boolean
      keep_structure:
        description: 是否保持本地文件结构
        type: boolean
      name:
        description: 配置名称
        type: string
      provider:
        description: OSS提供商：aliyun, tencent, qiniu
        type: string
      region:
        description: 区域
        type: string
      secret_key:
        description: 访问密钥Secret（不返回给前端）
        type: string
      sync_path:
        description: OSS同步路径前缀
        type: string
      updated_at:
        description: 修改时间
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 阿里云OSS文件管理系统 - Go语言实现的文件上传、下载、同步服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scinote Go Client API
  version: "1.0"
paths:
  /api/v1/files:
    get:
      description: 分页获取文件列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取文件列表
      tags:
      - 文件管理
  /api/v1/files/{id}:
    delete:
      description: 根据文件ID删除文件
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 文件不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除文件
      tags:
      - 文件管理
    get:
      description: 根据文件ID获取文件的详细信息
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文件信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 文件不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取文件信息
      tags:
      - 文件管理
    put:
      consumes:
      - multipart/form-data
      description: 根据文件ID更新文件内容
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      - description: 新的文件内容
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 文件不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新文件
      tags:
      - 文件管理
  /api/v1/files/{id}/download:
    get:
      description: 根据文件ID下载文件内容
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件内容
          schema:
            type: file
        "404":
          description: 文件不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 下载文件
      tags:
      - 文件管理
  /api/v1/files/search:
    get:
      description: 根据文件名搜索文件
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 搜索文件
      tags:
      - 文件管理
  /api/v1/files/stats:
    get:
      description: 获取系统中文件的统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 统计信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取文件统计信息
      tags:
      - 文件管理
  /api/v1/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传单个文件到服务器
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: 上传成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 上传文件
      tags:
      - 文件管理
  /oss/configs:
    get:
      consumes:
      - application/json
      description: 获取系统中所有已配置的OSS存储配置列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有OSS配置
      tags:
      - OSS配置管理
    post:
      consumes:
      - application/json
      description: 创建新的OSS存储配置，支持阿里云、腾讯云、七牛云等多种云存储服务
      parameters:
      - description: OSS配置信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/database.OSSConfig'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建OSS配置
      tags:
      - OSS配置管理
  /oss/configs/{id}:
    get:
      consumes:
      - application/json
      description: 根据配置ID获取指定的OSS配置详细信息
      parameters:
      - description: 配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/database.OSSConfig'
        "400":
          description: 配置ID无效
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 配置不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个OSS配置
      tags:
      - OSS配置管理
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"